pyproject.toml
[project]
name = "pythonproject11"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "self (>=2020.12.3,<2021.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "stripe (>=14.1.0,<15.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "decouple (>=0.0.7,<0.0.8)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
